Removes more dead code and standardizes non-lesson pages.
Before, I had the lesson and categories page as separate html, while the about page was markdown inserted into a layout file called pages, and later main pages. This was an unnecessary complication and was an artifact from when I was figuring out Jekyll.

Trying to clean up the code base.